:recycle: Refactor navigation components
Make them more granular and easy to scale
