--- src/navigation/helper.js
@@ -1,18 +1,0 @@
-/**
- * Since react-navigation doesn't support i18n
- * I've created this dummy function to help i18n scanner
- * understand about these titles.
- * We can remove this as soon as react-navigation supports i18n or
- * we change the router to another lib with i18n support.
- *
- * @param {String} str
- * @returns {String} same as the input string
- */
-export const t = str => str;
-
-export const headerStyle = {
-  backgroundColor: 'transparent',
-  overflow: 'hidden',
-  elevation: 1,
-  borderBottomWidth: 0,
-};
