+++ src/navigation/styles.js
@@ -1,0 +1,12 @@
+import { colors } from 'constants/styleGuide';
+
+export const navigationTabBarStyle = {
+  backgroundColor: colors.light.ultramarineBlue,
+  width: '90%',
+  bottom: 24,
+  borderRadius: 64,
+  minWidth: 300,
+  height: 72,
+  paddingTop: 24,
+  alignSelf: 'center',
+};
